SpeechBrain/requirements.txt
# core
torch==2.2.2             --extra-index-url https://download.pytorch.org/whl/cpu
torchaudio==2.2.2        --extra-index-url https://download.pytorch.org/whl/cpu
numpy==1.26.4
soundfile==0.12.1
librosa==0.10.2

# diarization
pyannote.audio==3.1.1
huggingface-hub==0.23.3

# ASR & fillers
openai-whisper==20231106
transformers==4.41.1
tokenizers==0.19.1          # pulled by transformers

# misc
scikit-learn==1.4.2
tqdm==4.66.4
ffmpeg-python==0.2.0